requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.1.0
fpdf2>=2.7.0
python-escpos>=3.0.0
pyusb>=1.2.1

